third_party/WebKit/LayoutTests/fast/css/invalidation/lazy-reattach-object.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<style>
    object + div { color: green; }
</style>
<object id="obj" data="data:text/html,FAIL"><div>Fallback</div></object>
<div>
    <div></div>
    <div></div>
    <div></div>
    <div></div>
</div>
<script>
description("Reattaching object should not trigger sibling forest invalidation.");

shouldBeDefined(window.internals);

// Force recalc
obj.offsetTop;

// Trigger reattach of fallback content.
obj.setAttribute("data", "data:text/html,PASS");

shouldBe("internals.updateStyleAndReturnAffectedElementCount()", "1");

</script>
