third_party/blink/web_tests/platform/generic/fast/css/invalidation/lazy-reattach-object-expected.txt

Reattaching object should not trigger sibling forest invalidation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS [object Internals] is defined.
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS successfullyParsed is true

TEST COMPLETE
